+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`1902cc2...main`][1902cc2...main].
+For a full diff see [`fd198f0...main`][fd198f0...main].
 
-[1902cc2...main]: https://github.com/ergebnis/rector-rules/compare/1902cc2...main
+### Added
+
+- Added `Rules\Arrays\SortAssociativeArrayByKeyRector` ([#81]), by [@localheinz]
+
+[fd198f0...main]: https://github.com/ergebnis/rector-rules/compare/fd198f0...main
+
+[#1]: https://github.com/ergebnis/rector-rules/pull/1
+
+[@localheinz]: https://github.com/localheinz